Added RandomSpike class. PulseGen messages do not seem to be functioning. Strange bug: new classes' fields are preceded by a dot when showing them.
git-svn-id: https://svn.code.sf.net/p/moose/code/moose/trunk@405 d1858a74-f715-0410-99c9-c66fe9cce3d6

+++ biophysics/PulseGen.h
@@ -55,7 +55,7 @@ class PulseGen
     //////////////////////////////////////////////////////////////////
     // Message dest functions.
     //////////////////////////////////////////////////////////////////
-    static void inputFunc(const Conn& c, int input);
+    static void inputFunc(const Conn& c, double input);
     static void setPulseLevel(const Conn& c, int pulseNo, double level);
     static void setPulseWidth(const Conn& c, int pulseNo, double width);
     static void setPulseDelay(const Conn& c, int pulseNo, double delay);
